refactor(layout): move away from mdsvex system (blogpost)
