+++ src/content/blogpost/project-euler/problem-001/index.en.svx
@@ -8,7 +8,6 @@
 		"math"
 	],
 	"lang": "en",
-	"layout": "blogpost",
 	"series": "project-euler",
 	"slug": "project-euler-1",
 	"title": "Project Euler #1",
