+++ src/content/blogpost/project-euler/problem-002/index.ru.svx
@@ -9,7 +9,6 @@
 		"math"
 	],
 	"lang": "ru",
-	"layout": "blogpost",
 	"series": "project-euler",
 	"slug": "проект-эйлера-2",
 	"title": "Проект Эйлера #2"
