+++ src/content/blogpost/quadtree/index.ru.svx
@@ -10,7 +10,6 @@
 		"computer science"
 	],
 	"lang": "ru",
-	"layout": "blogpost",
 	"slug": "дерево-квадрантов",
 	"title": "Дерево квадрантов",
 	"translation": {
