+++ src/content/blogpost/where-do-aperture-numbers-come-from/index.ru.svx
@@ -10,7 +10,6 @@
 		"f-number"
 	],
 	"lang": "ru",
-	"layout": "blogpost",
 	"slug": "откуда-берутся-числа-диафрагмы",
 	"title": "Откуда берутся числа диафрагмы?",
 	"translation": {
