+++ src/lib/layout/index.ts
@@ -1,4 +1,5 @@
 export { LayoutRoot } from "./root";
 export { LayoutPage } from "./page-header";
+export { default as BlogpostLayout } from "./blogpost/blogpost.layout.svelte";
 export { default as GalleryItemLayout } from "./gallery/gallery.layout.svelte";
 export { default as PageLayout } from "./page/page.layout.svelte";
